Renamable save file slot
- rom:/RDVHASH contains a string that will be the base for save files
  - i.e. if string is "foobar" saves are stored as "foobar0", "foobar1",
    "foobar2"
- split this and loading replacements.json into new "romfs.cpp/hpp"
  files
- added a header entry for nn::fs::GetEntryType
- added a constant for the offset of profile0 in string bank
- handles logging if replacements.json or RDVHASH is missing and
