+++ source/program/romfs.cpp
@@ -1,0 +1,177 @@
+#include "romfs.hpp"
+
+#include <nn.hpp>
+#include "lib.hpp"
+#include "cJSON.h"
+
+typedef struct
+{
+    u64 crc;
+    char const *replacement;
+} stringList;
+
+/* Function ptr to dread's crc function. */
+u64 (*crc64)(char const *str, u64 size) = NULL;
+
+stringList *g_stringList = NULL;
+size_t g_stringListSize = 0;
+
+CStrId* g_stringBank = NULL;
+
+/** Create a string instance */
+void (*create_string_instance) (CStrId* value, const char* str, uint len, u64 crc, bool storeInPool) = NULL;
+
+/* Takes in a pointer to string and if found in the list, is replaced with the desired string. */
+void replaceString(const char **str)
+{
+    /* Hash the string for quicker comparison. */
+    u64 crc = crc64(*str, strlen(*str));
+
+    /* Attempt to find matching hash in our list. */
+    for(size_t i = 0; i < g_stringListSize; i++)
+    {
+        /* If the string matches, replace the string. */
+        if(crc == g_stringList[i].crc)
+            *str = g_stringList[i].replacement;
+    }
+}
+
+/* Allocates a buffer and reads from the specified file. */
+void* odr::romfs::OpenAndReadFile(const char *path)
+{
+    nn::fs::DirectoryEntryType entryType;
+    nn::fs::GetEntryType(&entryType, path);
+    if (entryType != nn::fs::DirectoryEntryType_File) {
+        return NULL;
+    }
+
+    long int size;
+    nn::fs::FileHandle fileHandle;
+    nn::fs::OpenFile(&fileHandle, path, nn::fs::OpenMode_Read);
+    nn::fs::GetFileSize(&size, fileHandle);
+    u8 *fileBuf = (u8 *)malloc(size + 1);
+    nn::fs::ReadFile(fileHandle, 0, fileBuf, size);
+    nn::fs::CloseFile(fileHandle);
+    fileBuf[size] = '\0';
+    return fileBuf;
+}
+
+void populateStringReplacementList()
+{
+    /* Read contents of replacements.json in romfs. Allocates a heap buffer for the file contents and returns it. */
+    char *fileBuf = (char *)odr::romfs::OpenAndReadFile("rom:/replacements.json");
+    if (fileBuf == NULL) {
+        const char* msg = "[LogWarn/0] rom:/replacements.json does not exist!";
+        svcOutputDebugString(msg, strlen(msg));
+        return;
+    }
+
+    /* Parse json from file buffer. */
+    cJSON *json = cJSON_Parse(fileBuf);
+    if(json == NULL)
+        return;
+
+    /* Get the "replacements" object within the json. */
+    const cJSON *replacementList = cJSON_GetObjectItem(json, "replacements");
+
+    /* Get number of elements in the array, for later use. */
+    g_stringListSize = cJSON_GetArraySize(replacementList);
+
+    /* Allocate space for the list of strings to replace. */
+    g_stringList = (stringList *)malloc(g_stringListSize * sizeof(stringList));
+
+    size_t i = 0;
+    const cJSON *itemObject;
+
+    /* Iterate over array contents and extract strings. */
+    cJSON_ArrayForEach(itemObject, replacementList)
+    {
+        /* Extract the strings using the relevant method and add them to the list. */
+        if(cJSON_IsString(itemObject))
+        {
+            char *fileStr = cJSON_GetStringValue(itemObject);
+            char *replacementFileStr = (char *)malloc(strlen(fileStr) + strlen("rom:/") + 1);
+            strcpy(replacementFileStr, "rom:/");
+            replacementFileStr = strcat(replacementFileStr, fileStr);
+            g_stringList[i].crc = crc64(fileStr, strlen(fileStr));
+            g_stringList[i].replacement = replacementFileStr;
+        }
+        else if(cJSON_IsObject(itemObject))
+        {
+            char const *str = cJSON_GetItemName(itemObject);
+            g_stringList[i].crc = crc64(str, strlen(str));
+            g_stringList[i].replacement = cJSON_GetStringValue(itemObject->child);
+        }
+        i++;
+    }
+
+    /* Free the buffer allocated by OpenAndReadFile, since we're done with it. */
+    free(fileBuf);
+}
+
+void setSeedSaveProfile()
+{
+    char *seedHash = (char*)odr::romfs::OpenAndReadFile("rom:/RDVHASH");
+    if (seedHash == NULL) {
+        const char* msg = "[LogWarn/0] rom:/RDVHASH does not exist!";
+        svcOutputDebugString(msg, strlen(msg));
+        return;
+    }
+
+    int len = strlen(seedHash);
+    if (len > 0)
+    {
+        char slotName[256];
+        u64 crc;
+
+        sprintf(slotName, "%s0", seedHash);
+        crc = crc64(slotName, len+1);
+        create_string_instance(&g_stringBank[odr::romfs::STRINGBANK_PROFILE0], slotName, len+1, crc, true);
+
+        slotName[len] = '1';
+        crc = crc64(slotName, len+1);
+        create_string_instance(&g_stringBank[odr::romfs::STRINGBANK_PROFILE0+1], slotName, len+1, crc, true);
+
+        slotName[len] = '2';
+        crc = crc64(slotName, len+1);
+        create_string_instance(&g_stringBank[odr::romfs::STRINGBANK_PROFILE0+2], slotName, len+1, crc, true);
+    }
+    free(seedHash);
+    return;
+}
+
+HOOK_DEFINE_TRAMPOLINE(ForceRomfs) {
+    /* Define the callback for when the function is called. Don't forget to make it static and name it Callback. */
+    static void Callback(void *CFilePathStrIdOut, const char *path, u8 flags) {
+
+        /* Just in case the path is NULL, pass it down to the real implementation, since we don't support replacing NULL paths anyway. */
+        if(path == NULL)
+        {
+            Orig(CFilePathStrIdOut, path, flags);
+            return;
+        }
+
+        /* Replace string if we have it in our list before passing to the real implementation. */
+        replaceString(&path);
+        Orig(CFilePathStrIdOut, path, flags);
+    }
+};
+
+/* Hook romfs mounting. Pass the arguments down to the real implementation so romfs is mounted as normal. */
+/* Once romfs is mounted, we can read files from it in order to populate our string replacement list. */
+HOOK_DEFINE_TRAMPOLINE(RomMounted) {
+    static Result Callback(char const *path, void *romCache, unsigned long cacheSize) {
+        Result res = Orig(path, romCache, cacheSize);
+        populateStringReplacementList();
+        setSeedSaveProfile();
+        return res;
+    }
+};
+
+void odr::romfs::InstallHooks(functionOffsets* offsets) {
+    RomMounted::InstallAtFuncPtr(nn::fs::MountRom);
+    ForceRomfs::InstallAtOffset(offsets->CFilePathStrIdCtor);
+    crc64 = (u64 (*)(char const *, u64))exl::util::modules::GetTargetOffset(offsets->crc64);
+    create_string_instance = (void (*)(CStrId* value, const char* str, uint len, u64 crc, bool storeInPool)) exl::util::modules::GetTargetOffset(offsets->FindOrCreateStringInstance);
+    g_stringBank = (CStrId*)exl::util::modules::GetTargetOffset(offsets->StaticStringBank);
+}
